Separate packages in schedule.
In an effort to clean up the large schedule package, the favorite and
notification services were broken into separate packages.
