+++ src/main/java/com/animedetour/android/database/DetourDatabaseHelper.java
@@ -10,7 +10,7 @@ package com.animedetour.android.database;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
-import com.animedetour.android.schedule.Favorite;
+import com.animedetour.android.schedule.favorite.Favorite;
 import com.animedetour.api.guest.model.Category;
 import com.animedetour.api.guest.model.Guest;
 import com.animedetour.api.sched.api.model.Event;
