+++ src/main/java/com/animedetour/android/framework/dependencyinjection/module/ActivityModule.java
@@ -15,7 +15,7 @@ import com.animedetour.android.home.HomeFragment;
 import com.animedetour.android.main.MainActivity;
 import com.animedetour.android.schedule.DayFragment;
 import com.animedetour.android.schedule.EventActivity;
-import com.animedetour.android.schedule.FavoritesFragment;
+import com.animedetour.android.schedule.favorite.FavoritesFragment;
 import com.animedetour.android.schedule.ScheduleFragment;
 import com.animedetour.android.settings.SettingsFragment;
 import dagger.Module;
