+++ src/main/java/com/animedetour/android/framework/dependencyinjection/module/ApplicationModule.java
@@ -19,8 +19,8 @@ import com.android.volley.toolbox.Volley;
 import com.animedetour.android.BuildConfig;
 import com.animedetour.android.R;
 import com.animedetour.android.database.DataModule;
-import com.animedetour.android.schedule.EventNotificationManager;
-import com.animedetour.android.schedule.NotificationScheduler;
+import com.animedetour.android.schedule.notification.EventNotificationManager;
+import com.animedetour.android.schedule.notification.NotificationScheduler;
 import com.animedetour.android.volley.cache.LongImageCache;
 import com.animedetour.api.ApiModule;
 import com.google.android.gms.analytics.GoogleAnalytics;
