+++ src/main/java/com/animedetour/android/schedule/EventViewBinder.java
@@ -20,7 +20,7 @@ import com.inkapplications.android.widget.recyclerview.ViewClickListener;
  *
  * @author Maxwell Vandervelde (Max@MaxVandervelde.com)
  */
-class EventViewBinder implements ItemViewBinder<PanelView, Event>
+public class EventViewBinder implements ItemViewBinder<PanelView, Event>
 {
     /**
      * Context to be used when creating new panel views.
