+++ src/main/java/com/animedetour/android/schedule/favorite/Favorite.java
@@ -6,7 +6,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-package com.animedetour.android.schedule;
+package com.animedetour.android.schedule.favorite;
 
 import com.animedetour.api.sched.api.model.Event;
 import com.j256.ormlite.field.DatabaseField;
