+++ src/main/java/com/animedetour/android/schedule/notification/EventNotificationManager.java
@@ -6,7 +6,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-package com.animedetour.android.schedule;
+package com.animedetour.android.schedule.notification;
 
 import android.app.AlarmManager;
 import android.app.PendingIntent;
